A basic main function to make use of the 32-bit notification value.
Using a variation of notification sending to a task, where we set
desired bits of the 32-bit notification value.
The receiving task will look at the desired bits and if they're set, it
will perform a print to the serial terminal.

Still need to test the different actions after sending notification from
a task. But this idea could be used for setting flags that could be used
by a different task to do something in a certain condition.