Testing out the TaskNotify and TaskNotifyWait.
The action parameter of TaskNotify is important.
	You could overwrite previous sent notification by setting the action
parameter as: eSetValueWithOverwrite.

At the receiving task, it is good to clear the notification bits you're
testing in the task, once you've recieved a notification, i.e.
	xTaskNotifyWait(( (1<<3) | (1<<2) | (1<<1) |(1<<0)), 0,
&ulNotificationValue, portMAX_DELAY)

	In this, the received 32-bit notification value is stored in
ulNotificationValue and upon receiving bits 0,1,2,3 are cleared, because
we're going to check if these bits were set in the received
ulNotificationValue value.